requirements.txt
anyio==3.6.1
APScheduler==3.9.1
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==3.3.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
psutil==5.9.1
python-telegram-bot==20.0a1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
tornado==6.3.2
tzdata==2022.1
tzlocal==4.1
urllib3==2.1.0
